backend/requirements.txt
websockets==14.2
certifi
google-genai
aiohttp-cors
absl-py